requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mutagen==1.46.0
pyaes==1.6.1
pycryptodomex==3.17
Pyrogram==2.0.97
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
TgCrypto==1.2.5
websockets==10.4
yarl==1.8.2
yt-dlp==2023.1.6
